Add Resource_Info

+++ misc/accounting.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/bash
 #=============================================================================
+OOD_BASE_DIR=/var/www/ood/apps/sys/ondemand_fugaku
 ACCOUNTING_DIR=/system/ood/accounting
 LOCKFILE=${ACCOUNTING_DIR}/lockfile
 ELAPSED_TIME=${ACCOUNTING_DIR}/elapsed_time.txt
@@ -10,15 +11,22 @@ GROUP_TMP=${GROUP_DIR}/group.tmp
 DISK_TMP=${HOME_DIR}/disk.tmp
 INODE_TMP=${HOME_DIR}/inode.tmp
 FREE_QUEUE_TMP=${ACCOUNTING_DIR}/free_queue.tmp
-
 #=============================================================================
 ACCOUNTJ="/usr/local/bin/accountj"
 ACCOUNTD="/usr/local/bin/accountd"
+PJSTATA="/usr/local/bin/pjstata"
 CHKLOWPRIORITY="/usr/local/bin/chklowpriority"
 REMOTE_SSH=""
 [ "$HOSTNAME" != "fn06sv04" ] && REMOTE_SSH="ssh login"
+YEAR=$(date +%Y)
+MONTH=$(date +%m)
+PERIOD=1  # 前期（4月から9月）の場合は1、後期の場合は2
+FIRSTDAY=${YEAR}0401
+if [[ ${MONTH} -ge 10 || ${MONTH} -le 3 ]]; then
+  PERIOD=2
+  FIRSTDAY=${YEAR}1001
+fi
 #=============================================================================
-
 # ディスク容量が95%以上の場合、スクリプトを停止
 VAR=$(df / | tail -n 1 | awk '{print $5}' | sed -r 's/%//g')
 [ ${VAR} -ge 95 ] && exit 1
@@ -51,44 +59,84 @@ mkdir -p ${GROUP_DIR} ${HOME_DIR}
     # グループ毎のバジェットを取得
     for II in `cat ${GROUP_TMP}`; do
 	[ ${II} = "fujitsu" ] && continue
+
+        # 各グループのディレクトリを作成
+        DIR=${GROUP_DIR}/${II}
+        mkdir -p ${DIR}
+        chmod 750 ${DIR}
+        chown root:${II} ${DIR}
 	{
 	    # グループ名1つ1つに対してaccountj -gを実行し、グループ毎のバジェットを取得
-	    FILE=${GROUP_DIR}/${II}.resource
-	    su - ktool -c "${REMOTE_SSH} ${ACCOUNTJ} -g ${II} -c -h -r 1" | tr -d '"' | egrep '^SUBTHEME|^SUBTHEMEPERIOD' > ${FILE}
+	    FILE=${DIR}/resource.csv
+	    su - ktool -c "${REMOTE_SSH} ${ACCOUNTJ} -g ${II} -c -h -r 1 -e" | tr -d '"' | egrep '^SUBTHEME|^SUBTHEMEPERIOD|^EXCLUSIVEUSE' > ${FILE}
 	    
 	    if [ $? -ne 0 ]; then
-		rm ${LOCKFILE} ${GROUP_TMP}
-		exit 1
-	    fi
-	    chmod 640 ${FILE}
-	    chown root:${II} ${FILE}
-	} &
-	
-	{
-	    # グループ名1つ1つに対してaccountd -gを実行し、グループ毎のディスク使用量を取得
-	    # -Eをつけると、/dataと/shareの情報も出力される
-	    FILE=${GROUP_DIR}/${II}.disk
-	    su - ktool -c "${REMOTE_SSH} ${ACCOUNTD} -g ${II} -c -E" | tr -d '"' | tail -n +5 > ${FILE}
-	    
-	    if [ $? -ne 0 ]; then
-		rm ${LOCKFILE} ${GROUP_TMP}
-		exit 1
+	      rm ${LOCKFILE} ${GROUP_TMP}
+	      exit 1
 	    fi
 	    chmod 640 ${FILE}
 	    chown root:${II} ${FILE}
 	} &
 
 	{
-	    # グループ名1つ1つに対してaccountd -g -iを実行し、グループ毎のi-node使用量を取得
-	    FILE=${GROUP_DIR}/${II}.inode
-	    su - ktool -c "${REMOTE_SSH} ${ACCOUNTD} -g ${II} -i -c" | tr -d '"' | tail -n +4 > ${FILE}
-	    
+	    # グループ名1つ1つに対してaccountd -gを実行し、グループ毎のディスク使用量を取得
+	    # -Eをつけると、/dataと/shareの情報も出力される
+	    FILE=${DIR}/disk.csv
+	    su - ktool -c "${REMOTE_SSH} ${ACCOUNTD} -g ${II} -c -E" | tr -d '"' | tail -n +5 > ${FILE}
+
 	    if [ $? -ne 0 ]; then
-		rm ${LOCKFILE} ${GROUP_TMP}
-		exit 1
+	      rm ${LOCKFILE} ${GROUP_TMP}
+              exit 1
 	    fi
-	    chmod 640 ${FILE}
-	    chown root:${II} ${FILE}
+
+	    if [ -s ${FILE} ]; then
+	      chmod 640 ${FILE}
+	      chown root:${II} ${FILE}
+	    else
+	      rm -f ${FILE}
+	    fi
+
+	    # グループ名1つ1つに対してaccountd -g -iを実行し、グループ毎のi-node使用量を取得
+	    # 時間削減のために、ディスクを使用しているグループだけ以降の処理を行う
+	    if [ -s ${FILE} ]; then
+	      FILE=${DIR}/inode.csv
+	      su - ktool -c "${REMOTE_SSH} ${ACCOUNTD} -g ${II} -i -c" | tr -d '"' | tail -n +4 > ${FILE}
+	      
+	      if [ $? -ne 0 ]; then
+	        rm ${LOCKFILE} ${GROUP_TMP}
+	        exit 1
+	      fi
+	      
+              if [ -s ${FILE} ]; then
+                chmod 640 ${FILE}
+                chown root:${II} ${FILE}
+              else
+                rm -f ${FILE}
+              fi
+	    fi
+
+	    # グループ毎のバジェットを取得（前期 or 後期毎の各ユーザのリソース使用状況を取得）
+	    # 上と同様にディスクを使用しているグループだけ以降の処理を行うが、
+	    # rist-で始まるグループではディスクを利用していなくてもバジェットを利用しているので、それらは処理を行う
+	    if [[ -s ${FILE} || ${II} == rist-* ]]; then
+	      FILE=${DIR}/${YEAR}-${PERIOD}.csv
+
+	      # メモ：awkコマンドで下記のresource_info.rbを代替すると、CSVの要素の中にカンマがある場合に
+	      # 対応できないため、rubyのCSVモジュールを使っている
+	      su - ktool -c "${REMOTE_SSH} ${PJSTATA} -c -g ${II} -d ${FIRSTDAY}:" | ruby ${OOD_BASE_DIR}/misc/resource_info.rb > ${FILE}
+
+	      if [ $? -ne 0 ]; then
+                rm ${LOCKFILE} ${GROUP_TMP}
+                exit 1
+	      fi
+	      
+              if [ -s ${FILE} ]; then
+                chmod 640 ${FILE}
+                chown root:${II} ${FILE}
+              else
+                rm -f ${FILE}
+              fi
+	    fi
 	} &
 	wait
     done
@@ -171,9 +219,10 @@ mkdir -p ${GROUP_DIR} ${HOME_DIR}
     
     while read -r line; do
 	group=`echo $line | awk -F, '{print $3}'`
-	echo "$line" | awk -F, '{print $8}' > ${GROUP_DIR}/${group}.free_queue
-	chmod 640 ${GROUP_DIR}/${group}.free_queue
-	chown root:${group} ${GROUP_DIR}/${group}.free_queue
+	FILE=${GROUP_DIR}/${group}/free_queue.dat
+	echo "$line" | awk -F, '{print $8}' > ${FILE}
+	chmod 640 ${FILE}
+	chown root:${group} ${FILE}
     done < ${FREE_QUEUE_TMP}
 
     rm ${FREE_QUEUE_TMP}
